requirements.txt
fastapi[standard]
chromadb
httpx
python-dotenv